# docs/api/phase-4-openapi.yaml
openapi: 3.0.3
info:
  title: TeamFlow API
  version: 0.1.0
  description: Phase 4 project, task, label, and comment contract.
servers:
  - url: http://localhost:4000
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
  /organizations/{organizationId}/projects:
    get:
      summary: List organization projects
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - { name: search, in: query, schema: { type: string } }
        - {
            name: status,
            in: query,
            schema: { type: string, enum: [active, archived] },
          }
        - {
            name: sortBy,
            in: query,
            schema:
              { type: string, enum: [name, createdAt], default: createdAt },
          }
        - {
            name: sortOrder,
            in: query,
            schema: { type: string, enum: [asc, desc], default: desc },
          }
      responses:
        "200": { $ref: "#/components/responses/PaginatedResponse" }
    post:
      summary: Create a project
      description: Requires owner or admin membership.
      parameters: [{ $ref: "#/components/parameters/OrganizationId" }]
      requestBody: { $ref: "#/components/requestBodies/ProjectInput" }
      responses: { "201": { $ref: "#/components/responses/ProjectResponse" } }
  /organizations/{organizationId}/projects/{projectId}:
    get:
      summary: Get a project
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - $ref: "#/components/parameters/ProjectId"
      responses: { "200": { $ref: "#/components/responses/ProjectResponse" } }
    patch:
      summary: Update a project
      description: Requires owner or admin membership.
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - $ref: "#/components/parameters/ProjectId"
      requestBody: { $ref: "#/components/requestBodies/ProjectUpdate" }
      responses: { "200": { $ref: "#/components/responses/ProjectResponse" } }
    delete:
      summary: Delete a project and its tasks
      description: Requires owner or admin membership.
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - $ref: "#/components/parameters/ProjectId"
      responses: { "200": { $ref: "#/components/responses/DeletedResponse" } }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: teamflow_access_token }
    bearerAuth: { type: http, scheme: bearer }
  parameters:
    OrganizationId:
      {
        name: organizationId,
        in: path,
        required: true,
        schema: { type: string, format: uuid },
      }
    ProjectId:
      {
        name: projectId,
        in: path,
        required: true,
        schema: { type: string, format: uuid },
      }
...
    Page:
      {
        name: page,
        in: query,
        schema: { type: integer, minimum: 1, default: 1 },
      }
    PageSize:
      {
        name: pageSize,
        in: query,
        schema: { type: integer, minimum: 1, maximum: 100, default: 20 },
      }
  requestBodies:
    ProjectInput:
      required: true
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ProjectInput" } },
        }
    ProjectUpdate:
      required: true
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ProjectUpdate" } },
        }
...
  responses:
    ProjectResponse:
      {
        description: Project resource,
        content:
          {
            application/json:
              {
                schema:
                  {
                    type: object,
                    required: [project],
                    properties:
                      { project: { $ref: "#/components/schemas/Project" } },
                  },
              },
          },
      }
...
    PaginatedResponse:
      {
        description: Paginated list response,
        content:
          {
            application/json:
              {
                schema:
                  {
                    type: object,
                    required: [items, pagination],
                    properties:
                      {
                        items: { type: array, items: {} },
                        pagination: { $ref: "#/components/schemas/Pagination" },
                      },
                  },
              },
          },
      }
    DeletedResponse:
      {
        description: Deletion response,
        content:
          {
            application/json:
              {
                schema:
                  { type: object, properties: { deleted: { type: boolean } } },
              },
          },
      }
  schemas:
    ProjectInput:
      type: object
      required: [name, slug]
      properties:
        {
          name: { type: string, minLength: 2 },
          slug: { type: string, pattern: "^[a-z0-9-]{2,60}$" },
          description: { type: string, nullable: true },
        }
    ProjectUpdate:
      allOf: [{ $ref: "#/components/schemas/ProjectInput" }]
      properties: { status: { type: string, enum: [active, archived] } }
...
    Project:
      {
        type: object,
        properties:
          {
            id: { type: string, format: uuid },
            organizationId: { type: string, format: uuid },
            name: { type: string },
            slug: { type: string },
            status: { type: string },
            description: { type: string, nullable: true },
          },
      }
...
    Pagination:
      type: object
      required: [page, pageSize, total]
      properties:
        {
          page: { type: integer },
          pageSize: { type: integer },
          total: { type: integer },
        }